clean_up_files splitting up main
now we have a profile that allows for a separation of concerns. Main no
longer need be concerned with the apollo client retrieval of profile
data
