+++ src/Main/Main.js
@@ -1,24 +1,13 @@
 /* eslint no-undef: 0 */
 import React from 'react';
 import PropTypes from 'prop-types';
-import $ from 'jquery';
-import { Link as RouterLink } from 'react-router-dom';
+// import { Link as RouterLink } from 'react-router-dom';
 // MAterial UI generic
 import clsx from 'clsx';
 import { makeStyles, useTheme } from '@material-ui/core/styles';
-import Drawer from '@material-ui/core/Drawer';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import List from '@material-ui/core/List';
-import Typography from '@material-ui/core/Typography';
-import Divider from '@material-ui/core/Divider';
-import IconButton from '@material-ui/core/IconButton';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemIcon from '@material-ui/core/ListItemIcon';
-import ListItemText from '@material-ui/core/ListItemText';
-import { Avatar } from '@material-ui/core';
-// icons
+import { Drawer, CssBaseline, AppBar, Toolbar, List, Typography, Divider,
+  IconButton, ListItem, ListItemText, ListItemIcon }  from '@material-ui/core';
+// tree shaking material-ui icons is problematic
 import MenuIcon from '@material-ui/icons/Menu';
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
 import ChevronRightIcon from '@material-ui/icons/ChevronRight';
@@ -28,12 +17,8 @@ import EditIcon from '@material-ui/icons/Edit';
 import NotificationsIcon from '@material-ui/icons/Notifications';
 import ExitToAppIcon from '@material-ui/icons/ExitToApp';
 const drawerWidth = 240;
-// Apollo
-import { gql } from 'apollo-boost';
-import { useQuery } from '@apollo/react-hooks';
-import * as queries from '../graphql/queries';
-
-const meQuery = gql(queries.users.getMe.graphql);
+// custom imports
+import { Profile } from './components';
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -106,166 +91,125 @@ const Main = props => {
   const handleDrawerClose = () => {
     setOpen(false);
   };
-
-  // DATA BINDING
-  const { loading, error, data } = useQuery(meQuery);
-  if (error) {
-    return <div>Error</div>;
-  }
-
-  if (loading) {
-    return (
-      <div className="App">
-        <h2>Loading...</h2>
-      </div>
-    );
-  }
-  if (data) {
-    console.log('user is', data.me);
-    // if (data.user.length > 0) {
-    if (!$.isEmptyObject(data.me)) {
-      return (
-        <div className={classes.root}>
-          <CssBaseline />
-          <AppBar
-            className={clsx(classes.appBar, {
-              [classes.appBarShift]: open,
-            })}
-            position="fixed"
+  return (
+    <div className={classes.root}>
+      <CssBaseline />
+      <AppBar
+        className={clsx(classes.appBar, {
+          [classes.appBarShift]: open,
+        })}
+        position="fixed"
+      >
+        <Toolbar>
+          <IconButton
+            aria-label="open drawer"
+            className={clsx(classes.menuButton, open && classes.hide)}
+            color="inherit"
+            edge="start"
+            onClick={handleDrawerOpen}
           >
-            <Toolbar>
-              <IconButton
-                aria-label="open drawer"
-                className={clsx(classes.menuButton, open && classes.hide)}
-                color="inherit"
-                edge="start"
-                onClick={handleDrawerOpen}
-              >
-                <MenuIcon />
-              </IconButton>
-              <Typography
-                noWrap
-                variant="h6"
-              >
-                { /* text for our header */ }
-              </Typography>
-            </Toolbar>
-          </AppBar>
-          <Drawer
-            anchor="left"
-            classes={{
-              paper: classes.drawerPaper,
-            }}
-            className={classes.drawer}
-            open={open}
-            variant="persistent"
+            <MenuIcon />
+          </IconButton>
+          <Typography
+            noWrap
+            variant="h6"
           >
-            <div className={classes.drawerHeader}>
-              <IconButton onClick={handleDrawerClose}>
-                {theme.direction === 'ltr' ? <ChevronLeftIcon /> : <ChevronRightIcon />}
-              </IconButton>
-              <Divider />
-              <div
-                className={clsx(classes.root)}
-              >
-                <Avatar
-                  alt="Person"
-                  className={classes.avatar}
-                  component={RouterLink}
-                  src={data.me.profileImage}
-                  to="/settings"
-                />
-                <Typography
-                  className={data.me.initials}
-                  variant="h4"
-                >
-                  {data.me.phone}
-                </Typography>
-                <Typography variant="body2">
-                  {`${data.me.first} ${data.me.last}`}
-                  <br />
-                  {data.me.zipCode}
-                </Typography>
-              </div>
-            </div>
-            <Divider />
-            <List>
-              <ListItem>
-                <ListItemIcon>
-                  <AccountIcon />
-                </ListItemIcon>
-                <ListItemText primary="View my profile" />
-              </ListItem>
-              <ListItem>
-                <ListItemIcon>
-                  <FamilyIcon />
-                </ListItemIcon>
-                <ListItemText primary="View my family" />
-              </ListItem>
-              <ListItem>
-                <ListItemIcon>
-                  <EditIcon />
-                </ListItemIcon>
-                <ListItemText primary="Edit my profile" />
-              </ListItem>
-              <ListItem>
-                <ListItemIcon>
-                  <NotificationsIcon />
-                </ListItemIcon>
-                <ListItemText primary="Messages" />
-              </ListItem>
-              <ListItem>
-                <ListItemIcon>
-                  <ExitToAppIcon />
-                </ListItemIcon>
-                <ListItemText primary="Setup" />
-              </ListItem>
-              <ListItem>
-                <ListItemIcon>
-                  <ExitToAppIcon />
-                </ListItemIcon>
-                <ListItemText primary="Sign out" />
-              </ListItem>
-            </List>
-
-            <Divider />
-          </Drawer>
-          <main
-            className={clsx(classes.content, {
-              [classes.contentShift]: open,
-            })}
-          >
-            <div className={classes.drawerHeader} />
-            <Typography paragraph>
-              DISPLAY DATA FOR SELECTION OF DRAWER HERE ???
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt
-              ut labore et dolore magna aliqua. Rhoncus dolor purus non enim praesent elementum
-              facilisis leo vel. Risus at ultrices mi tempus imperdiet. Semper risus in hendrerit
-              gravida rutrum quisque non tellus. Convallis convallis tellus id interdum velit laoreet id
-              donec ultrices. Odio morbi quis commodo odio aenean sed adipiscing. Amet nisl suscipit
-              adipiscing bibendum est ultricies integer quis. Cursus euismod quis viverra nibh cras.
-              Metus vulputate eu scelerisque felis imperdiet proin fermentum leo. Mauris commodo quis
-              imperdiet massa tincidunt. Cras tincidunt lobortis feugiat vivamus at augue. At augue eget
-              arcu dictum varius duis at consectetur lorem. Velit sed ullamcorper morbi tincidunt. Lorem
-              donec massa sapien faucibus et molestie ac.
-            </Typography>
-            <Typography paragraph>
-              Consequat mauris nunc congue nisi vitae suscipit. Fringilla est ullamcorper eget nulla
-              facilisi etiam dignissim diam. Pulvinar elementum integer enim neque volutpat ac
-              tincidunt. Ornare suspendisse sed nisi lacus sed viverra tellus. Purus sit amet volutpat
-              consequat mauris. Elementum eu facilisis sed odio morbi. Euismod lacinia at quis risus sed
-              vulputate odio. Morbi tincidunt ornare massa eget egestas purus viverra accumsan in. In
-              hendrerit gravida rutrum quisque non tellus orci ac. Pellentesque nec nam aliquam sem et
-              tortor. Habitant morbi tristique senectus et. Adipiscing elit duis tristique sollicitudin
-              nibh sit. Ornare aenean euismod elementum nisi quis eleifend. Commodo viverra maecenas
-              accumsan lacus vel facilisis. Nulla posuere sollicitudin aliquam ultrices sagittis orci a.
-            </Typography>
-            {children}
-          </main>
+            { /* text for our header */ }
+          </Typography>
+        </Toolbar>
+      </AppBar>
+      <Drawer
+        anchor="left"
+        classes={{
+          paper: classes.drawerPaper,
+        }}
+        className={classes.drawer}
+        open={open}
+        variant="persistent"
+      >
+        <div className={classes.drawerHeader}>
+          <IconButton onClick={handleDrawerClose}>
+            {theme.direction === 'ltr' ? <ChevronLeftIcon /> : <ChevronRightIcon />}
+          </IconButton>
+          <Divider />
+          <Profile />
         </div>
-      );
-    }
-  }
+        <Divider />
+        <List>
+          <ListItem>
+            <ListItemIcon>
+              <AccountIcon />
+            </ListItemIcon>
+            <ListItemText primary="View my profile" />
+          </ListItem>
+          <ListItem>
+            <ListItemIcon>
+              <FamilyIcon />
+            </ListItemIcon>
+            <ListItemText primary="View my family" />
+          </ListItem>
+          <ListItem>
+            <ListItemIcon>
+              <EditIcon />
+            </ListItemIcon>
+            <ListItemText primary="Edit my profile" />
+          </ListItem>
+          <ListItem>
+            <ListItemIcon>
+              <NotificationsIcon />
+            </ListItemIcon>
+            <ListItemText primary="Messages" />
+          </ListItem>
+          <ListItem>
+            <ListItemIcon>
+              <ExitToAppIcon />
+            </ListItemIcon>
+            <ListItemText primary="Setup" />
+          </ListItem>
+          <ListItem>
+            <ListItemIcon>
+              <ExitToAppIcon />
+            </ListItemIcon>
+            <ListItemText primary="Sign out" />
+          </ListItem>
+        </List>
+
+        <Divider />
+      </Drawer>
+      <main
+        className={clsx(classes.content, {
+          [classes.contentShift]: open,
+        })}
+      >
+        <div className={classes.drawerHeader} />
+        <Typography paragraph>
+          DISPLAY DATA FOR SELECTION OF DRAWER HERE ???
+          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt
+          ut labore et dolore magna aliqua. Rhoncus dolor purus non enim praesent elementum
+          facilisis leo vel. Risus at ultrices mi tempus imperdiet. Semper risus in hendrerit
+          gravida rutrum quisque non tellus. Convallis convallis tellus id interdum velit laoreet id
+          donec ultrices. Odio morbi quis commodo odio aenean sed adipiscing. Amet nisl suscipit
+          adipiscing bibendum est ultricies integer quis. Cursus euismod quis viverra nibh cras.
+          Metus vulputate eu scelerisque felis imperdiet proin fermentum leo. Mauris commodo quis
+          imperdiet massa tincidunt. Cras tincidunt lobortis feugiat vivamus at augue. At augue eget
+          arcu dictum varius duis at consectetur lorem. Velit sed ullamcorper morbi tincidunt. Lorem
+          donec massa sapien faucibus et molestie ac.
+        </Typography>
+        <Typography paragraph>
+          Consequat mauris nunc congue nisi vitae suscipit. Fringilla est ullamcorper eget nulla
+          facilisi etiam dignissim diam. Pulvinar elementum integer enim neque volutpat ac
+          tincidunt. Ornare suspendisse sed nisi lacus sed viverra tellus. Purus sit amet volutpat
+          consequat mauris. Elementum eu facilisis sed odio morbi. Euismod lacinia at quis risus sed
+          vulputate odio. Morbi tincidunt ornare massa eget egestas purus viverra accumsan in. In
+          hendrerit gravida rutrum quisque non tellus orci ac. Pellentesque nec nam aliquam sem et
+          tortor. Habitant morbi tristique senectus et. Adipiscing elit duis tristique sollicitudin
+          nibh sit. Ornare aenean euismod elementum nisi quis eleifend. Commodo viverra maecenas
+          accumsan lacus vel facilisis. Nulla posuere sollicitudin aliquam ultrices sagittis orci a.
+        </Typography>
+        {children}
+      </main>
+    </div>
+  );
 };
 
 Main.propTypes = {
