+++ src/Main/components/index.js
@@ -1,0 +1,2 @@
+export { default as Profile } from './Profile';
+// export { default as SidebarNav } from './SidebarNav';
